macos:  Give me the first 19 Calculator Button Files
nineteen because

    eleven at:  0 1 2 3 4 5 6 7 8 9 .
    plus the original gang of four at:  + - * /
    and then the final four at:  , over π clear

like now the Fibonacci Numbers stack up at

    Clear  1 , 2  Over Over +  Over Over +  ...

Finder > View > As List (⌘2) will show you that,
    after you press those ten Buttons inside Finder > View > As Icons (⌘1)

    for W in  clear  1 comma 2  over over plus  over over plus; do
        (cd macos/ && ./$W.command)
    done

+++ bin/byopyvm.py
@@ -47,22 +47,25 @@ examples:
 
   # Maths
 
-  = math.pi  # echo 3.141592653589793 >3.142~
-  = 2 *  # echo 2 >2~ && rm -f 3.142~ 2~ && echo 6.283185307179586 >6.283~
-  = .  # cat 6.283~ && rm -f 6.283~
+  = math.pi  # echo 3.141592653589793 >3.142
+  = 2 *  # echo 2 >2 && rm -f 3.142 2 && echo 6.283185307179586 >6.283
+  = .  # cat 6.283 && rm -f 6.283
 
   = math.pi 2 * .  # all at once
 """
 
 
+import collections
 import json
 import math
+import os
 import pathlib
 import pdb
 import re
 import shlex
 import subprocess
 import sys
+import traceback
 
 import byotools as byo
 
@@ -79,8 +82,6 @@ def main():
     main.parms = parms
 
     collapse_star_parms(parms)
-
-    func_by_verb = form_func_by_verb()
 
     patchdoc = """
 
@@ -101,7 +102,17 @@ def main():
     byo.exit_if_testdoc()  # byopyvm.py
     byo.exit_if_argdoc()  # byopyvm.py --help
 
+    # Run well
+
+    parms_run(parms)
+
+
+def parms_run(parms):
+    """Run the Parms = Read a Word, Evaluate the Word, Print the Result, Loop"""
+
     # Take each word, one at a time
+
+    func_by_verb = form_func_by_verb()
 
     while parms:
 
@@ -110,6 +121,8 @@ def main():
         shverb = parms[0]
         if shverb not in func_by_verb:
             shverb = taker_from_word(word=shverb)
+
+        assert shverb in func_by_verb, (shverb, parms)
 
         func = func_by_verb[shverb]
         func(parms)
@@ -122,8 +135,10 @@ def collapse_star_parms(parms):
 
     shline = "ls"
     shshline = "bash -c {!r}".format(shline)
-    argv = shlex.split(shshline)
 
+    sys.stdout.flush()
+    sys.stderr.flush()
+    argv = shlex.split(shshline)
     run = subprocess.run(argv, stdin=subprocess.PIPE, stdout=subprocess.PIPE)
     stdout = run.stdout.decode()
     lines = stdout.splitlines()
@@ -197,6 +212,17 @@ def form_name_by_char():
     d["."] = "dot"
     d["/"] = "slash"
 
+    d["0"] = "zero"  # Python accepts Digital Digits in the Name, but not to start with
+    d["1"] = "one"
+    d["2"] = "two"
+    d["3"] = "three"
+    d["4"] = "four"
+    d["5"] = "five"
+    d["6"] = "six"
+    d["7"] = "seven"
+    d["8"] = "eight"
+    d["9"] = "nine"
+
     d[":"] = "colon"
     d[";"] = "semi"
     # d["<"]
@@ -241,10 +267,15 @@ def form_func_by_verb():
     """Declare the Pipe Filter Abbreviations"""
 
     func_by_verb = dict(
+        buttonfile=do_buttonfile,
         dot=do_dot,
         dotted_name=do_dotted_name,
         int=do_int,
         ls=do_ls,
+        minus=do_minus,
+        name=do_name,
+        over=do_push_y,
+        plus=do_plus,
         slash=do_slash,
         star=do_star,
     )
@@ -255,16 +286,15 @@ def form_func_by_verb():
 def do_dot(parms):
     """Pop X but print its Value"""
 
-    x = stack_pop(promise="cat {} && ")
+    x = stack_pop(1, default=None, promise="cat {} && ")
 
     print(x)
 
 
 def do_dotted_name(parms):
-    """Push the Value of a Dotted Name"""
+    """Call a Dotted Name, else push its Value"""
 
-    evalled = eval(parms[0])
-    stack_push(evalled)
+    do_name(parms)
 
 
 def do_int(parms):
@@ -282,22 +312,59 @@ def do_ls(parms):
 
     shshline = "bash -c {!r}".format(shline)
     argv = shlex.split(shshline)
+
+    sys.stdout.flush()
+    sys.stderr.flush()
     subprocess.run(argv, stdin=subprocess.PIPE, check=True)
 
 
-def do_slash(parms):
-    """Push Y / X"""
+def do_minus(parms):
+    """Push Y - X in place of Y X"""
 
-    (_x, _y) = stack_pop(2)
-    x = _x / _y
+    (_y, _x) = stack_pop(2, default=0)
+    x = _y - _x
+    stack_push(x)
+
+
+def do_plus(parms):
+    """Push Y + X in place of Y X"""
+
+    (_y, _x) = stack_pop(2, default=0)
+    x = _y + _x
+    stack_push(x)
+
+
+def do_name(parms):
+    """Call a Name, else push its Value"""
+
+    evalled = eval(parms[0])
+
+    if not isinstance(evalled, collections.abc.Callable):
+        stack_push(evalled)
+    else:
+        evalled(parms)
+
+
+def do_push_y(parms):
+    """Push Y"""
+
+    (_y, _x) = stack_peek(2)
+    stack_push(_y)
+
+
+def do_slash(parms):
+    """Push Y / X in place of Y X"""
+
+    (_y, _x) = stack_pop(2, default=1)
+    x = _y / _x
     stack_push(x)
 
 
 def do_star(parms):
-    """Push Y * X"""
+    """Push Y * X in place of Y X"""
 
-    (_x, _y) = stack_pop(2)
-    x = _x * _y
+    (_y, _x) = stack_pop(2, default=1)
+    x = _y * _x
     stack_push(x)
 
 
@@ -306,49 +373,77 @@ def do_star(parms):
 #
 
 
-def stack_pop(depth=1, promise=""):
-    """Peek and remove some of the Values most recently pushed"""
+def stack_pop(depth, default=None, promise=""):
+    """Peek and eval and remove some of the Values most recently pushed"""
 
-    pairs = stack_pop_pairs(depth, promise=promise)
+    peeks = stack_peek(depth, default=default)
+
+    _ = stack_pairs_pop(depth, promise=promise)
+
+    return peeks
+
+
+def stack_peek(depth, default):
+    """Peek and eval some of the Values most recently pushed"""
+
+    assert depth >= 1
+
+    pairs = stack_pairs_peek(depth, default=default)  # peek, not pop
     values = list(_[-1] for _ in pairs)
 
-    pops = list()
+    peeks = list()
     for value in values:
-        pop = json.loads(value)
-        pops.append(pop)
 
-    if len(pops) == 1:
-        one_pop = pops[-1]
+        peek = default
+        if value is not None:
+            try:
+                peek = json.loads(value)
+            except json.JSONDecodeError:
+                peek = default
 
-        return one_pop
+        peeks.append(peek)
 
-    return pops
+    if len(peeks) == 1:
+        one_peek = peeks[-1]
+
+        return one_peek
+
+    return peeks
 
 
-def stack_pop_pairs(depth=1, promise=""):
+def stack_pairs_pop(depth, default=None, promise=""):
     """Peek and remove some of the Key-Value Pairs most recently pushed"""
 
-    pairs = stack_peek_pairs(depth)
+    assert depth >= 1
+
+    pairs = stack_pairs_peek(depth, default=default)
 
     paths = list(_[0] for _ in pairs)
-    shpaths = " ".join(byo.shlex_dquote(_) for _ in paths)
-    shline = "rm -f {}".format(shpaths)
+    shpaths = " ".join(byo.shlex_dquote(_) for _ in paths if _ is not None)
+    if shpaths:
+        shline = "rm -f {}".format(shpaths)
 
-    sys.stderr.write("+ {}{}\n".format(promise.format(shpaths), shline))
+        sys.stderr.write("+ {}{}\n".format(promise.format(shpaths), shline))
 
-    argv = shlex.split(shline)
-    subprocess.run(argv, stdin=subprocess.PIPE, stdout=subprocess.PIPE, check=True)
+        sys.stdout.flush()
+        sys.stderr.flush()
+        argv = shlex.split(shline)
+        subprocess.run(argv, stdin=subprocess.PIPE, stdout=subprocess.PIPE, check=True)
 
     return pairs
 
 
-def stack_peek_pairs(depth=1):
+def stack_pairs_peek(depth=1, default=None):
     """Peek at some of the Key-Value Pairs most recently pushed"""
+
+    assert depth >= 1
 
     #
 
     shline = "ls -1rt"
 
+    sys.stdout.flush()
+    sys.stderr.flush()
     argv = shlex.split(shline)
     run = subprocess.run(
         argv, stdin=subprocess.PIPE, stdout=subprocess.PIPE, check=True
@@ -356,17 +451,33 @@ def stack_peek_pairs(depth=1):
     stdout = run.stdout.decode()
     lines = stdout.splitlines()
 
+    filenames = lines
+
     #
 
     pairs = list()
 
-    filenames = lines[-depth:]
     for filename in filenames:
         path = pathlib.Path(filename)
-        chars = path.read_text()
+        if path.is_file():
 
-        pair = (str(path), chars)
-        pairs.append(pair)
+            chars = path.read_text()
+            try:
+                _ = json.loads(chars)
+            except json.JSONDecodeError:  # todo: stack more types, not just Json Chars
+
+                continue
+
+            pair = (str(path), chars)
+            pairs.append(pair)
+
+    pairs = pairs[-depth:]  # todo: stop evalling the discarded Depths of Stack
+
+    #
+
+    none_key_pair = (None, default)
+    while len(pairs) < depth:
+        pairs.append(none_key_pair)
 
     #
 
@@ -376,18 +487,200 @@ def stack_peek_pairs(depth=1):
 def stack_push(value):
     """Push the Json Chars of a Value, into a new Autonamed File"""
 
-    key = "{}~".format(round(value, 3))
+    if isinstance(value, float):
+        key = "{}".format(round(value, 3))
+    else:
+        key = str(value)
+
     stack_push_key_value(key, value=value)
 
 
 def stack_push_key_value(key, value):
     """Push the Json Chars of a Value, into a fresh File"""
 
-    sys.stderr.write("+ echo {} >{}\n".format(value, key))
+    path = pathlib.Path(key)
+    shpath = byo.shlex_dquote(str(path))
+
+    shvalue = byo.shlex_dquote(str(value))
+
+    alt_path = find_alt_path(path)
+    alt_shpath = byo.shlex_dquote(str(alt_path))
+
+    mv_shline = "mv -i {} {}".format(shpath, alt_shpath)
+    echo_shline = "echo {} >{}".format(shvalue, shpath)
+
+    if not path.exists():
+        sys.stderr.write("+ {}\n".format(echo_shline))
+    else:
+        sys.stderr.write("+ {} && {}\n".format(mv_shline, echo_shline))
+
+        os.rename(path, dst=alt_path)
 
     chars = json.dumps(value)
     with open(key, "w") as writing:
         writing.write(chars)
+
+
+def find_alt_path(path):
+
+    alt_path = pathlib.Path("{}~".format(path))  # the 0th Alt
+
+    index = 1
+    while alt_path.exists():
+        alt_path = pathlib.Path("{}~{}~".format(path, index))
+
+        index += 1
+
+    return alt_path
+
+
+#
+# Take Double-Click's of Dot-Command ButtonFile's
+#
+
+
+def do_buttonfile(parms):
+    """Take one Double-Click of a Dot-Command ButtonFile"""
+
+    try:
+
+        do_buttonfile_word(parms)
+
+    except Exception:
+        sys.stderr.write("\n")
+
+        traceback.print_exc()
+
+        sys.stderr.write("Press ⌃D TTY EOF to quit\n")
+        sys.stdin.read()
+
+        raise
+
+
+def do_buttonfile_word(parms):
+    """Run the Name of a Dot-Command ButtonFile, without its Ext, as a Word"""
+
+    assert parms
+
+    # Take the Name, without its Ext, as a Word
+
+    main_file = parms.pop(1)
+
+    basename = os.path.basename(main_file)
+    (root, ext) = os.path.splitext(basename)
+    _ = ext
+
+    # Run the Word
+
+    word = root
+    if word in "0123456789":
+        entry_write_char(parms=[word])
+    elif word == "E":  # todo:  add "+:-" to toggle the Sign of Exp else of Mantissa
+        entry_write_char(parms=["E".lower()])
+    elif word == "clear":
+        entries_clear()
+    elif word == "dot":
+        entry_write_char(parms=["."])
+    else:
+        entry = entry_close_if_open()
+        if word == "comma":
+            run_button_comma(entry)
+
+        elif word == "\N{Greek Small Letter Pi}":  # π
+            do_dotted_name(parms=["math.pi"])
+
+        else:
+
+            run_button_word(parms, word=word)
+
+
+def run_button_word(parms, word):
+    """Run the Parms, but first close the last Entry, if needed"""
+
+    entry_close_if_open()
+
+    parms = [word]
+    parms_run(parms)
+
+
+def entries_clear():
+    """Clear this Dir of the Git Clone, and the Dir's it contains"""
+
+    shline = "git clean -f {}".format(byo.shlex_dquote(os.getcwd()))
+    sys.stderr.write("+ {}\n".format(shline))
+
+    sys.stdout.flush()
+    sys.stderr.flush()
+    subprocess.run(shlex.split(shline))
+
+
+def run_button_comma(entry):
+    """Run '0' and ',' if no Entry preceded Comma, else do nothing but return"""
+
+    if entry is None:
+        entry_write_char(parms=["0"])
+        entry_close_if_open()
+
+
+def entry_write_char(parms):
+    """Push the first Char, or append a later Char, of an Int or Float Literal"""
+
+    word = parms.pop(0)
+
+    entry = pop_entry(default="")
+    entry += word
+    entry += "_"
+
+    stack_push(entry)
+
+
+def entry_close_if_open():
+    """Push the first Char, or append a later Char, of an Int or Float Literal"""
+
+    entry = pop_entry()
+    if entry is not None:
+
+        try:
+            evalled = int(entry)
+        except ValueError:
+            evalled = float(entry)
+
+        stack_push(evalled)
+
+    return entry  # not Eval'led
+
+
+def pop_entry(default=None):
+    """Pop the collected Chars and return them, else return the chosen Default"""
+
+    entry = peek_entry(default=None)
+    if entry is None:
+
+        return default
+
+    _ = stack_pop(1)
+
+    return entry
+
+
+def peek_entry(default):
+    """Peek the collected Chars and return them, else return the chosen Default"""
+
+    entry = default
+
+    pair = stack_pairs_peek(1)[-1]
+    if pair is not None:
+        (key, value) = pair
+
+        if key.endswith("_"):
+            key_json = json.dumps(key)
+            if key_json == value:
+                evalled = json.loads(value)
+                if re.match("^[-+.0-9][-+.0-9Ee]*_$", string=evalled):
+
+                    entry = byo.str_removesuffix(evalled, suffix="_")
+
+    return entry
 
 
 #
